Ticket [https://trac.osgeo.org/fdo/ticket/761]: FDO API: Use INT_MAX instead of LONG_MAX for FdoInt32. LONG_MAX == LLONG_MAX on Linux x64
git-svn-id: http://svn.osgeo.org/fdo/trunk@6927 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Fdo/Unmanaged/Src/Common/Io/MemoryStream.cpp
@@ -133,8 +133,8 @@ void FdoIoMemoryStream::Write( FdoIoStream* stream, FdoSize count )
     // Write the data to this internal and its following buffers, until all bytes
     // have been written.
     while ( lCount > 0 ) {
-        // Our buffer array can't have more than LONG_MAX members
-        if ( bufNum >= LONG_MAX )
+        // Our buffer array can't have more than INT_MAX members
+        if ( bufNum >= INT_MAX )
             throw FdoException::Create(
                 FdoException::NLSGetMessage(
                     FDO_NLSID(FDO_34_MEMORYSTREAMOVERFLOW),
